# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "TaskMaster API - Enterprise multi-project task management system",
    "version": "1.0.0",
    "title": "TaskMaster API",
    "contact": {
      "name": "TaskMaster Support",
      "email": "support@taskmaster.dev"
    }
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter 'Bearer' followed by your JWT token"
    }
  },
  "paths": {
    "/health": {
      "get": {
        "tags": ["Health"],
        "summary": "Health Check",
        "description": "Check if the server is running",
        "responses": {
          "200": {
            "description": "Server is healthy",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "ok"
                },
                "time": {
                  "type": "string",
                  "example": "2025-08-23T12:00:00Z"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User Login",
        "description": "Login with email and password to get JWT token",
        "parameters": [
          {
            "in": "body",
            "name": "credentials",
            "description": "Login credentials",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["email", "password"],
              "properties": {
                "email": {
                  "type": "string",
                  "example": "admin@taskmaster.dev"
                },
                "password": {
                  "type": "string",
                  "example": "admin123"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful",
            "schema": {
              "type": "object",
              "properties": {
                "access_token": {
                  "type": "string"
                },
                "refresh_token": {
                  "type": "string"
                },
                "token_type": {
                  "type": "string",
                  "example": "Bearer"
                },
                "expires_in": {
                  "type": "integer",
                  "example": 86400
                },
                "user": {
                  "type": "object",
                  "properties": {
                    "id": {"type": "string"},
                    "email": {"type": "string"},
                    "username": {"type": "string"},
                    "role": {"type": "string"}
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User Registration",
        "description": "Register a new user account",
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "description": "User registration data",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["email", "username", "password", "role"],
              "properties": {
                "email": {
                  "type": "string",
                  "example": "user@example.com"
                },
                "username": {
                  "type": "string",
                  "example": "testuser"
                },
                "password": {
                  "type": "string",
                  "example": "password123"
                },
                "role": {
                  "type": "string",
                  "enum": ["admin", "manager", "developer", "tester", "viewer"],
                  "example": "developer"
                },
                "first_name": {
                  "type": "string",
                  "example": "John"
                },
                "last_name": {
                  "type": "string", 
                  "example": "Doe"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "409": {
            "description": "User already exists"
          }
        }
      }
    },
    "/api/v1/users/me": {
      "get": {
        "tags": ["Users"],
        "summary": "Get Current User",
        "description": "Get information about the currently authenticated user",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "User information",
            "schema": {
              "type": "object",
              "properties": {
                "id": {"type": "string"},
                "email": {"type": "string"},
                "username": {"type": "string"},
                "role": {"type": "string"},
                "is_active": {"type": "boolean"},
                "created_at": {"type": "string"},
                "updated_at": {"type": "string"}
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/projects": {
      "get": {
        "tags": ["Projects"],
        "summary": "List Projects",
        "description": "Get list of projects",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "List of projects"
          }
        }
      },
      "post": {
        "tags": ["Projects"],
        "summary": "Create Project",
        "description": "Create a new project",
        "security": [{"BearerAuth": []}],
        "responses": {
          "201": {
            "description": "Project created"
          }
        }
      }
    },
    "/api/v1/tasks": {
      "get": {
        "tags": ["Tasks"],
        "summary": "List Tasks",
        "description": "Get list of tasks",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "List of tasks"
          }
        }
      },
      "post": {
        "tags": ["Tasks"],
        "summary": "Create Task", 
        "description": "Create a new task",
        "security": [{"BearerAuth": []}],
        "responses": {
          "201": {
            "description": "Task created"
          }
        }
      }
    }
  }
}